# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/users/index": {
            "get": {
                "tags": [
                    "User Management"
                ],
                "summary": "Get list of users",
                "operationId": "152ae101a32f0f352efaf73fd3aee2ab",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search user by name, phone and email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "food"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for paginated results",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "0:deactivated, 1:activated",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "perpage",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/users/detail": {
            "get": {
                "tags": [
                    "User Management"
                ],
                "summary": "Get detail of User",
                "operationId": "64cf30b8edcd2b618b9899131d1b01b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/users/change-status": {
            "post": {
                "tags": [
                    "User Management"
                ],
                "summary": "Change the status of a user",
                "operationId": "f0e4e56c775c0e9914b208e0d788c9e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "User ID to change status",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/users/delete": {
            "delete": {
                "tags": [
                    "User Management"
                ],
                "summary": "Delete a User",
                "operationId": "b699309f6a7abc8284be7f4e59f65dc1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "User ID is deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}